build: 게시판 관련 빌드를 admin으로 통합
infoBoard관련 Controller, Dao interface, Dao class....등을 admin으로 통합하고 관련
Command를 변경, 기타 수정사항 적용(기능 변경은 없음)
