+++ ycHackers/src/main/java/com/ych/pjt/command/InfoBoardPageListCommand.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.ui.Model;
 
-import com.ych.pjt.dao.InfoDao;
+import com.ych.pjt.dao.AdminDao;
 import com.ych.pjt.dto.InfoBoardDto;
 import com.ych.pjt.util.Constant;
 
@@ -14,10 +14,10 @@ public class InfoBoardPageListCommand implements IYchCommand {
 
 	@Override
 	public void execute(HttpServletRequest req, Model model) {
-		InfoDao iDao = Constant.iDao;
+		AdminDao adDao = Constant.adDao;
 		String pageNum = req.getParameter("pageNo");
 		System.out.println("infoBoard page List command: "+pageNum);
-		ArrayList<InfoBoardDto> dtos = iDao.infoPageList(pageNum);
+		ArrayList<InfoBoardDto> dtos = adDao.infoPageList(pageNum);
 		model.addAttribute("listContent", dtos);
 	}
