+++ ycHackers/src/main/java/com/ych/pjt/controller/AdminController.java
@@ -10,10 +10,12 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.ych.pjt.command.ChangeGradeCommand;
 import com.ych.pjt.command.IYchCommand;
+import com.ych.pjt.command.InfoBoardHomeCommand;
+import com.ych.pjt.command.InfoBoardManageCommand;
+import com.ych.pjt.command.InfoBoardPageListCommand;
 import com.ych.pjt.command.UserGradeCommand;
 import com.ych.pjt.command.UserSearchCommand;
 import com.ych.pjt.dao.AdminDao;
-import com.ych.pjt.dao.ResDao;
 import com.ych.pjt.util.Constant;
 
 @Controller
@@ -27,15 +29,43 @@ public class AdminController {
 		Constant.adDao=adDao;
 	}
 	
-	@RequestMapping("/admin")
+	@RequestMapping("/admin")//admin page
 	public String admin(HttpServletRequest req,Model model) {
 		System.out.println("admin page request");
 		//공지 게시판 최근 5개 호출 command
 		//page 정보 관련 data 호출 command
 		return "adminPage";
 	}	
-	//사용되는 DB를 중점으로 해야할지 사용자를 중점으로 할지 정해야할득
-	//user관련은 user Controller?
+	/*공지게시판 관련 request*/
+	@RequestMapping("/infoBoardHome")
+	public String infoBoardHome(HttpServletRequest req, Model model) {
+		System.out.println("infoBoard show from Home request");
+		com = new InfoBoardHomeCommand();
+		com.execute(req, model);
+		return "infoBoardHome";
+	}
+	@RequestMapping("/infoBoardManage")//관리자
+	public String infoBoardManage(HttpServletRequest req, Model model) {
+		System.out.println("infoBoardManage request");
+		com = new InfoBoardManageCommand();
+		com.execute(req, model);
+		return "infoBoardManage";
+	}
+	@RequestMapping("/infoPList")
+	public String infoPList(HttpServletRequest req, Model model) {
+		String pageNum = req.getParameter("pageNo");
+		System.out.println("infoBoard page request: "+pageNum);
+		com=new InfoBoardPageListCommand();
+		com.execute(req, model);
+		return "infoBoardPList";
+	}
+	@RequestMapping("/infoBoardWriteForm")
+	public String infoBoardWriteForm() {
+		System.out.println("infoBoardWriteForm request");
+		return "infoBoardWriteForm";
+	}
+
+	/*회원 권한 관련 request*/
 	@RequestMapping("/userGrade")
 	public String userGrade(HttpServletRequest req, Model model) {
 		System.out.println("userGrade");
