+++ ycHackers/src/main/java/com/ych/pjt/controller/CafeController.java
@@ -1,10 +1,8 @@
 package com.ych.pjt.controller;
 
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,12 +15,6 @@ import com.ych.pjt.util.Constant;
 public class CafeController {
 	private IYchCommand com;
 	
-	private BCryptPasswordEncoder passwordEncoder;
-	@Autowired
-	public void setPasswordEncoder(BCryptPasswordEncoder passwordEncoder) {
-		this.passwordEncoder = passwordEncoder;
-		Constant.passwordEncoder = passwordEncoder;
-	}
 	private CafeDao cDao;
 	@Autowired
 	public void setDao(CafeDao cDao) {
