--- ycHackers/src/main/java/com/ych/pjt/dao/IInfoDao.java
@@ -1,11 +1,0 @@
-package com.ych.pjt.dao;
-
-import java.util.ArrayList;
-
-import com.ych.pjt.dto.InfoBoardDto;
-
-public interface IInfoDao {
-	public ArrayList<InfoBoardDto> infoBoardHome();
-	public ArrayList<InfoBoardDto> infoBoardManage();
-	public ArrayList<InfoBoardDto> infoPageList(String pageNum);
-}
